+++ filen-cli/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Docker builds
+- `export-api-key` command for use with unmanaged Rclone
 
 ## 0.2.2 - 2025-12-23